Chore: rename initUserKeyAndID to initUserData & code style
